+++ src/main/java/se/plilja/imcollect/MapBasedImmutableList.java
@@ -4,13 +4,13 @@ import java.util.Iterator;
 
 // TODO: 06/03/16 Replace with finger trees
 final class MapBasedImmutableList<T> implements ImmutableList<T> {
-    private final WeightBalancedTreeMap<Long, T> map;
+    private final WeightBalancedTreeMap<Integer, T> map;
 
     public MapBasedImmutableList() {
-        this(new WeightBalancedTreeMap<>(Long::compare));
+        this(new WeightBalancedTreeMap<>(Integer::compare));
     }
 
-    private MapBasedImmutableList(WeightBalancedTreeMap<Long, T> map) {
+    private MapBasedImmutableList(WeightBalancedTreeMap<Integer, T> map) {
         this.map = map;
     }
 
@@ -32,11 +32,11 @@ final class MapBasedImmutableList<T> implements ImmutableList<T> {
         if (value == null) {
             return this;
         }
-        for (long i = 0; i < size(); i++) {
+        for (int i = 0; i < size(); i++) {
             T otherValue = map.get(i);
             if (otherValue.equals(value)) {
-                WeightBalancedTreeMap<Long, T> newMap = new WeightBalancedTreeMap<>(Long::compare);
-                for (long j = 0; j < size(); j++) {
+                WeightBalancedTreeMap<Integer, T> newMap = new WeightBalancedTreeMap<>(Integer::compare);
+                for (int j = 0; j < size(); j++) {
                     if (j != i) {
                         newMap = newMap.put(newMap.size(), map.get(j));
                     }
@@ -49,7 +49,7 @@ final class MapBasedImmutableList<T> implements ImmutableList<T> {
     }
 
     @Override
-    public long size() {
+    public int size() {
         return map.size();
     }
 
@@ -58,7 +58,7 @@ final class MapBasedImmutableList<T> implements ImmutableList<T> {
         if (val == null) {
             return false;
         }
-        for (long i = 0; i < size(); i++) {
+        for (int i = 0; i < size(); i++) {
             if (map.get(i).equals(val)) {
                 return true;
             }
@@ -72,7 +72,7 @@ final class MapBasedImmutableList<T> implements ImmutableList<T> {
     }
 
     @Override
-    public T get(long idx) {
+    public T get(int idx) {
         assertValidIndex(idx);
         return map.get(idx);
     }
@@ -84,14 +84,14 @@ final class MapBasedImmutableList<T> implements ImmutableList<T> {
     }
 
     @Override
-    public MapBasedImmutableList<T> set(long idx, T value) {
+    public MapBasedImmutableList<T> set(int idx, T value) {
         assertValidIndex(idx);
         return new MapBasedImmutableList<>(map.put(idx, value));
     }
 
     @Override
-    public long indexOf(T value) {
-        for (long i = 0; i < size(); i++) {
+    public int indexOf(T value) {
+        for (int i = 0; i < size(); i++) {
             if (map.get(i).equals(value)) {
                 return i;
             }
@@ -100,8 +100,8 @@ final class MapBasedImmutableList<T> implements ImmutableList<T> {
     }
 
     @Override
-    public long lastIndexOf(T value) {
-        for (long i = size() - 1; i >= 0; i--) {
+    public int lastIndexOf(T value) {
+        for (int i = size() - 1; i >= 0; i--) {
             if (map.get(i).equals(value)) {
                 return i;
             }
